+++ testing/test_terminal.py
@@ -600,7 +600,10 @@ def test_getreportopt():
 def test_terminalreporter_reportopt_addopts(testdir):
     testdir.makeini("[pytest]\naddopts=-rs")
     testdir.makepyfile("""
-        def pytest_funcarg__tr(request):
+        import pytest
+
+        @pytest.fixture
+        def tr(request):
             tr = request.config.pluginmanager.getplugin("terminalreporter")
             return tr
         def test_opt(tr):
@@ -614,7 +617,10 @@ def test_terminalreporter_reportopt_addopts(testdir):
 
 def test_tbstyle_short(testdir):
     p = testdir.makepyfile("""
-        def pytest_funcarg__arg(request):
+        import pytest
+
+        @pytest.fixture
+        def arg(request):
             return 42
         def test_opt(arg):
             x = 0
@@ -625,7 +631,7 @@ def test_tbstyle_short(testdir):
     assert 'arg = 42' not in s
     assert 'x = 0' not in s
     result.stdout.fnmatch_lines([
-        "*%s:5*" % p.basename,
+        "*%s:8*" % p.basename,
         "    assert x",
         "E   assert*",
     ])